+++ js/core/app.js
@@ -27,7 +27,7 @@ import { initializePerformanceController, markPerformance, measurePerformance, s
 import { initializeSaveEditModeSync } from './saveEditModeSync.js';
 import { initializeLayoutStabilityController } from '../platform/shell/layoutStabilityController.js';
 
-const APP_VERSION = '1.3.3-dev.3';
+const APP_VERSION = '1.3.3-dev.4';
 initializeLayoutStabilityController();
 initializePerformanceController({ appVersion: APP_VERSION });
 const appInitStartMark = markPerformance('app:init:start', { appVersion: APP_VERSION });